Proof of concept: error focus
Possible alternative way to do the linking to a specific form input to gain the focus, from the error summary, in a very decoupled way.

Only gotcha so far I've found is we have a dynamic form object (`caution_type_form`) which shows different radios based on a previous question (`under_age`), so with this solution we can't know where to link in a static way.

There are possible ways around this, but perhaps better to settle with the default behaviour (same we have now, link to the fieldset, not a specific radio) while we work on introducing more form objects as it might well be this solution can't handle all scenarios.