Update markup where needed
We have to introduce the `data-module` attribute in the `<details>` tags (we only have one of these in this service).

Prefix the error summary data-module with `govuk`.

Use the new markup for the `start now` button where the arrow is an SVG.

Any other markup changes that we may detect will be done as part of separate PRs to keep this one small and easy to review.