Cleanup of unused javascript code
We took most of this code from the C100 service, but in this service we don't need to track many of the things we were tracking on C100, and there is no need to have all this unused javscript code around.

If we need this in the future, we will add only what we need.