Add missing `data-module` to buttons
Ensure all the buttons or links that look and behave like a button have the `data-module="govuk-button"` attribute.

DRY a bit this by creating a helper method (`#link_button`) so it is easier to make changes in the future, just in one place.

Remove pointless `<div class="form-submit">` wrapper around some of these buttons.