Bump to `gov_uk_date_fields` gem (inputmode)
The recommended way to use the numeric inputs is to be of type `text` but with pattern and inputmode attributes.

The dates gem has been updated to produce this new markup [0] following the reports of issues with previous markup [1].

Also updated the `conviction_length` step to produce an input of type `text`, instead of number, and add the `inputmode` attribute.

[0] https://github.com/ministryofjustice/gov_uk_date_fields/releases/tag/v4.1.0
[1] https://github.com/alphagov/govuk-frontend/issues/1449