Fixes scenario 9 by legal
Unfortunately we did not caught this before, at first I thought it was
an issue with the user input order (and kinda is).

In the given scenario, YRO is compared with Community Order
the variable _spent_date_ upon comparison changes to the spent date of
Community Order, although our next date comparison uses the unmodified
YRO _spent_date_without_relevant_orders_ which backs to 2014, this means
that it wouldn't take into account the new computed spent date
(from previous iteration).

The code modification is kinda ugly, but the main purpose is that
we should use the _spent_date_ variable for date comparison if this
variable has been modified from the initial assignment, this means that
there was a carry over from a previous iteration.

Story: https://trello.com/c/r3BsvxQC