Remove "show_results" parameter
We used this before when we had both singles and multiples, because singles didn't have CYA. This was always a nasty workaround so I'm glad we can remove this finally.

Updated the decision trees as all journeys now end in the CYA instead of the results page as before for singles.