Remove unused views
For singles we used "variants" to render one of the many different views depending if it was a caution, conviction, spent, not spent, etc.

For multiples we are not using this approach so far so better to clean up.

In doing this I realised we didn't add the feedback partial to the new multiples results page, so I added it.