Refactor the CYA and Results presenters
This will make the presenters more flexible because now, each one will have an associated view (`to_partial_path`) but also a scope, that can be `check_your_answers` or `results` because we will want to present the information in a different way depending on which page we are.

Apart from this, some other cleanup and refactor of locales.