More design tweaks to CYA/results
Just adding the nice boxes around each of the proceedings (cautions or convictions) as I think this really makes a difference to clarify each of them.

Although the results page needs to change dramatically, I've added the box there too for consistency as it looks nicer.

This is following some of the latest changes to the prototype, so still WIP but starting to take form.