Remove reference to non endorsed FPN
Back in PR #466 we have agreed to only have endorsed FPN's in the 
service - this is due to non endorsed FPN not being considered a 
conviction. 

Now that we have the possibility to calculate multiple convictions, we 
no longer support non endorsed FPN's as they are not a conviction.

This is more of a follow up cleanup on #466