Remove conviction date from SpentDatePanel
I've reconsidered this, I think it does not read as well as the previous 
text. 
The other reason is that cautions do not have a conviction date, they 
have "known_date" and this causes more code to check the kind of the 
offence. I don't think it's necessary, as we know this information from 
the table underneath.

So I've removed the conviction date from the text to avoid more 
complications and repetition.

Happy to revisit this later on if we think it's necessary.

Story: https://trello.com/c/QmITv0Rd