Refactor and DRY some code in preparation for DBS
Now that we have exposed an `#spent?` method in the `ResultsVariant` we can simplify some code.

Also, instead of repeating the `variant` or `tense` code to know if a given date is in the past or not, we just pass to the `SpentTag` and to the (future) `DbsVisibility` the variant as they don't need to know the specific spent_date or do any calculation with it.