Functionality to remove sentences in CYA
This adds functionality for the user to be able to remove sentences
in the Check Your Answers page.

When a sentence is deleted, it's crucial to know wether there was other sentences
in the same report. Specially when the deleted sentence was the last one to be created,
since, until now, the session relied purely on the last `disclosure_check`, if that is deleted
we must find the Report id in order to know if there was other sentences present in order to send
the user to CYA or to the first page of the service (and restart their journey).

Story: https://trello.com/c/D6Goc7YV