Add validation to caution date
For conditional cautions, the "known_date" (caution date) must be before the conditions end date.

This so far has been fine but with the `change` functionality in the CYA, the user might edit a conditional caution date and introduce a wrong date without noticing.

We might need to revisit other places where this can happen in convictions for the same reason, but for now I'm focusing on cautions as per PR #519.