Some attributes can jump back to CYA
Some attributes can be edited without having to go through the following steps. This way after changing something, we can take the user back to the CYA page quickly.

This is done by marking these attributes with the `?next_step=cya` query param.

After editing something and clicking the `continue` button, if the `next_step` was set, the decision tree logic will be skipped and jump straight to the path specified in the next step (it is a hidden tag in the form).

If the attribute is the last one in a journey, there is no need to add this query param as it is redundant because the next step is always the CYA in these cases.
For example for cautions, `conditional_end_date` is the last attribute in the journey and the next step is always the CYA anyways.