Introduce days to sentence length
We have received feedback that sometimes courts give the sentence in days,
which means that users would have to convert those days into one of our options,
this is not ideal as it could prone users to introduce decimal numbers (which are not allowed).

So this commit allows users to introduce days, and internally we convert those days into months.

I've removed the conversion constants and replaced the calculations with ActiveSupport own methods.

The reason the specs have been changeds is that ActiveSupport calculates seconds per month,
so there's was a tiny discrepancy with the previous implementation, but nothing drastic (at least not in short periods of time)

ActiveSupport::Duration::SECONDS_PER_MONTH

1.month.in_seconds
=> 2629746

Story: https://trello.com/c/DqoWLnHu