Remove nginx reverse proxy
We were serving the static assets from a different nginx container in kubernetes. But in order to simplify the setup for the migration to the new EKS live cluster, we are removing this extra overhead as it doesn't really makes too much sense for a service like this, with almost no assets other than a few stylesheets and fonts.

The k8s deploy definition will also be updated:
https://github.com/ministryofjustice/disclosure-checker-deploy/pull/44