Implement cookies permission banner - Part 2
Ticket: https://trello.com/c/i5IitID9

Follow-up to PR #564.

Show the cookies permission banner on page load, and hook the accept/reject buttons to the controller.

Add a bit of javascript to allow to close the banner without reloading the page.

This follows the guidelines and design stated in the design system:
https://design-system.service.gov.uk/components/cookie-banner/

In the next PR we will implement the cookies preferences in the cookies page and add cucumber tests.