Add ECR images cleaner workflow
We come from #572 and #573 (now closed).

This is part of the work I've been doing for DEX Uhura (CT-3300) to come up with a reusable solution to delete old ECR images without risking deleting active images.

The repo for the action is here:
https://github.com/ministryofjustice/ecr-images-cleaner-action

This workflow will run on a schedule every Sunday at 8am, as well as it is possible to run it manually (once it is merged to `main`).

It will delete images older than 30 days, that are not part of the replica set history for production namespace, and keeping a buffer of 50 images no matter how old they are.